Convert a matrix to a FLPS data

Convert a matrix to a FLPS data

makeFLPSdata(
  inp_data,
  outcome,
  group,
  covariate,
  lv_model,
  lv_type,
  custom = F,
  ...
)

Arguments

inp_data

A matrix or a data frame

outcome

A character indicating the name of an outcome variable

group

A character indicating the name of a treatment/control group variable

covariate

A character indicating the names of covariates variables

lv_model

A description of the latent variable model, which is similar to the lavaan model syntax.

lv_type

A character indicating the type of latent variable models

...

Additional arguments for latent variable models information (e.g., nclass = 2).

custom_data

A list. should be provided with custome_stan

custom_stan

A string. should be provided with custome_data

Value

a S3 class for a corresponding measurement model