Generate a matrix style data for simulation

makeInpDat is a function for generating a data based on the given information.

makeInpData(
  N,
  R2Y,
  omega,
  tau0,
  tau1,
  lambda,
  R2eta,
  linear,
  nsec,
  lvmodel,
  lvinfo
)

Arguments

N

a numeric indicating sample size.

R2Y

a numeric indicating predictive power of covariates.

omega

a numeric indicating the relationship between eta_T and Y_C.

tau0

a numeric indicating the treatment eff when eta_T=0.

tau1

a numeric indicating the relationship between eta_T and Y_T-Y_C; (for now, single level (not multilevel), linear, normal).

lambda

a numeric indicating the mean of Worked problems/person.

R2eta

a numeric indicating Predictive power of latent variable (extent to which covariates predict eta).

nsec

a numeric indicating the number of maximum sections given to students.

lvmodel

a character specifying a type of latent variable model.

Value

a list containing latent variable model information, true eta, and a matrix containing the data for running FLPS.

Examples

sdat <- makeInpData(
  N = 100,
  R2Y = 0.2,
  omega = 0.2,
  tau0 = 0.13,
  tau1 = -0.06,
  lambda = 10,
  R2eta = 0.5,
  nsec = 10,
  linear = T,
  lvmodel = "sem",
  lvinfo  = 1
)
#> Error: $ operator is invalid for atomic vectors