Generate Fully Latent Principal Stratification data for simulation

makeSimData is a function for generating a data based on the given information.

makeSimData(
  N,
  R2Y,
  R2eta,
  omega,
  tau0,
  tau1,
  linear,
  ydist,
  lambda,
  nitem,
  nfac,
  lvmodel,
  fcovmat,
  item.missing = T,
  misspec = F,
  cov.res = 0,
  relsize = 0.6
)

Arguments

N

a numeric indicating sample size.

R2Y

a numeric indicating predictive power of covariates.

R2eta

a numeric indicating Predictive power of latent variable

omega

a numeric indicating the size of effect of latent factor on the outcome.

tau0

a numeric indicating the size of difference in the outcome between the treatment and the control.

tau1

a numeric indicating the principal effect

linear

a logical

ydist

a character

lambda

a numeric indicating the mean of Worked problems/person. (extent to which covariates predict eta).

nitem

a numeric indicating the number of maximum measurement items given to students.

nfac

a numeric indicating the number of latent factors

lvmodel

a character specifying a type of latent variable model.

fcovmat

a matrix indicating the variance-covariance matrix of latent factors when nfac > 1

item.missing

a logical to make the measurement item data missing for the control group. Default is TRUE

misspec

a logical to allow cross-loadings across latent factors when nfac > 1.

cov.res

a logical to allow for residual correlations (only for CFA model).

relsize

a numeric indicating the degree to which the latent factor explain the variances of continous items (only for CFA model).

Value

a list containing all the data related to population values and running FLPS.

Examples

sdat <- makeSimData(
N       = 200,  # sample size
R2Y     = 0.2,  # r^2 of outcome
R2eta   = 0.5,  # r^2 of eta by one covariates
omega   = 0.2,  # the effect of eta
tau0    = 0.13, # direct effect
tau1    = -0.06,# interaction effect between Z and eta
lambda  = 0.8,  # the proportion of administered items
nitem    = 10,   # the total number of items
nfac    = 1,    # the number of latent factors
lvmodel = '2pl' )