Conduct fully latent principal stratification

Conduct fully latent principal stratification

runFLPS(
  inp_data = NULL,
  complied_stan = NULL,
  outcome = NULL,
  trt = NULL,
  covariate = NULL,
  lv_model = NULL,
  lv_type = NULL,
  multilevel = FALSE,
  priors_input = NULL,
  stan_options = list(),
  ...
)

Arguments

inp_data

A matrix or a data frame

complied_stan

A Rds file. should be generated by modelBuilder.

outcome

A character indicating the name of an outcome variable

trt

A character indicating the name of a treatment/control group variable

covariate

A character indicating the names of covariate variables

lv_model

A description of the latent variable model, which is similar to the lavaan model syntax.

=~ : Specify the association between factors and indicators (e.g., F1 =~ v1 + v2 + v3).

+ : Specify a series of indicators

lv_type

A character indicating the type of latent variable models

multilevel

A logical specifying multilevel structure.

priors_input

A list of priors. Otherwise, the default priors are used (N(0, 5). It takes three parameter names including tau0, tau1, and omega, which are the difference between groups, the principal effects, and the effect of latent factors on the outcome. If added, the length of tau1 and omega must be matched with the number of factors. Examples of How to specify priors as follows:

list(tau0 = c(0, 1), tau1 = c(0.5, 1)) : The first element is the mean and the second is the variance of normal priors.

list(tau1 = list(c(0.5, 1), c(-0.4, 1)) : If there's two factors.

stan_options

A list containing [rstan::stan()] options, using 'name = value'.

...

Additional arguments for latent variable models information (e.g., nclass = 2).

Value

an object of class flps which contains a stanfit object.

call

argument calls

inp_data

A given data frame

flps_model

a Stan syntax used in [rstan::stan()]

flps_data

a list of data used in [rstan::stan()]

flps_fit

stanfit

time

a numeric of timing

See also

Examples

# \donttest{
inp_data <- flps::makeInpData(
  N       = 200,
  R2Y     = 0.2,
  R2eta   = 0.5,
  omega   = 0.2,
  tau0    = 0.23,
  tau1    = -0.16,
  betaL   = 0.1,
  betaY   = 0.2,
  lambda  = 0.8,
  nitem    = 10,
  nfac    = 1,
  lvmodel = 'rasch' )

res <- runFLPS(
   inp_data = inp_data,
   outcome = "Y",
   trt = "Z",
   covariate = c("X1"),
   lv_type = "rasch",
   lv_model = "F =~ v1 + v2 + v3 + v4 + v5 + v6 + v7 + v8 + v9 + v10",
   stan_options = list(iter = 1000, warmup = 500, cores = 1, chains = 2)
   )
#> Compiling Stan code...
#> 
#> SAMPLING FOR MODEL 'anon_model' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 0.00059 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 5.9 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:   1 / 1000 [  0%]  (Warmup)
#> Chain 1: Iteration: 100 / 1000 [ 10%]  (Warmup)
#> Chain 1: Iteration: 200 / 1000 [ 20%]  (Warmup)
#> Chain 1: Iteration: 300 / 1000 [ 30%]  (Warmup)
#> Chain 1: Iteration: 400 / 1000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 500 / 1000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 501 / 1000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 600 / 1000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 700 / 1000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 800 / 1000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 900 / 1000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 1000 / 1000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 9.336 seconds (Warm-up)
#> Chain 1:                7.191 seconds (Sampling)
#> Chain 1:                16.527 seconds (Total)
#> Chain 1: 
#> Warning: There were 8 divergent transitions after warmup. See
#> https://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup
#> to find out why this is a problem and how to eliminate them.
#> Warning: There were 1 chains where the estimated Bayesian Fraction of Missing Information was low. See
#> https://mc-stan.org/misc/warnings.html#bfmi-low
#> Warning: Examine the pairs() plot to diagnose sampling problems
#> Warning: The largest R-hat is NA, indicating chains have not mixed.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#r-hat
#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#bulk-ess
#> Warning: Tail Effective Samples Size (ESS) is too low, indicating posterior variances and tail quantiles may be unreliable.
#> Running the chains for more iterations may help. See
#> https://mc-stan.org/misc/warnings.html#tail-ess
# }